test-refactor(cargo): Renamed test classes to match tested classes. Removed redundant test.
