+++ detectable/src/test/java/com/synopsys/integration/detectable/detectables/cargo/transform/CargoLockPackageDataTransformerTest.java
@@ -1,67 +1,59 @@
 package com.synopsys.integration.detectable.detectables.cargo.transform;
 
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import com.synopsys.integration.bdio.graph.DependencyGraph;
-import com.synopsys.integration.bdio.graph.builder.MissingExternalIdException;
-import com.synopsys.integration.bdio.model.Forge;
-import com.synopsys.integration.detectable.detectable.exception.DetectableException;
+import com.synopsys.integration.detectable.detectables.cargo.data.CargoLockPackageData;
 import com.synopsys.integration.detectable.detectables.cargo.model.CargoLockPackage;
-import com.synopsys.integration.detectable.util.CycleDetectedException;
+import com.synopsys.integration.detectable.detectables.cargo.parse.CargoDependencyLineParser;
 import com.synopsys.integration.detectable.util.NameOptionalVersion;
-import com.synopsys.integration.detectable.util.graph.NameVersionGraphAssert;
 import com.synopsys.integration.util.NameVersion;
 
-public class CargoLockPackageDataTransformerTest {
+class CargoLockPackageDataTransformerTest {
 
     @Test
-    public void testParsesNamesAndVersionsSimple() throws DetectableException, MissingExternalIdException, CycleDetectedException {
-        List<CargoLockPackage> input = new ArrayList<>();
-        input.add(createPackage("test1", "1.0.0"));
-        input.add(createPackage("test2", "2.0.0"));
-        CargoLockPackageTransformer cargoLockPackageTransformer = new CargoLockPackageTransformer();
-        DependencyGraph graph = cargoLockPackageTransformer.transformToGraph(input);
+    void transform() {
+        CargoDependencyLineParser cargoDependencyLineParser = new CargoDependencyLineParser();
+        CargoLockPackageDataTransformer transformer = new CargoLockPackageDataTransformer(cargoDependencyLineParser);
 
-        NameVersionGraphAssert graphAssert = new NameVersionGraphAssert(Forge.CRATES, graph);
-        graphAssert.hasRootSize(2);
-        graphAssert.hasRootDependency("test1", "1.0.0");
-        graphAssert.hasRootDependency("test2", "2.0.0");
+        List<String> dependencies = Arrays.asList("dep1", "dep2 2.0.0", "dep3 3.0.0 (registry+https://some-registry-url");
+        CargoLockPackageData cargoLockPackageData = new CargoLockPackageData("some-name", "some-version", null, null, dependencies);
+        CargoLockPackage cargoLockPackage = transformer.transform(cargoLockPackageData);
+
+        assertEquals(new NameVersion("some-name", "some-version"), cargoLockPackage.getPackageNameVersion());
+        assertEquals(3, cargoLockPackage.getDependencies().size());
+
+        NameOptionalVersion dep1Actual = cargoLockPackage.getDependencies().get(0);
+        assertEquals("dep1", dep1Actual.getName());
+        assertFalse(dep1Actual.getVersion().isPresent());
+
+        NameOptionalVersion dep2Actual = cargoLockPackage.getDependencies().get(1);
+        assertEquals("dep2", dep2Actual.getName());
+        assertTrue(dep2Actual.getVersion().isPresent());
+        assertEquals("2.0.0", dep2Actual.getVersion().get());
+
+        NameOptionalVersion dep3Actual = cargoLockPackage.getDependencies().get(2);
+        assertEquals("dep3", dep3Actual.getName());
+        assertTrue(dep3Actual.getVersion().isPresent());
+        assertEquals("3.0.0", dep3Actual.getVersion().get());
     }
 
     @Test
-    public void multiplePackageDefinitionsTest() {
-        List<CargoLockPackage> input = new ArrayList<>();
-        input.add(createPackage("child1", "version1"));
-        input.add(createPackage("child1", "version2"));
-        input.add(createPackage("parent1", "anything", new NameOptionalVersion("child1")));
-        CargoLockPackageTransformer cargoLockPackageTransformer = new CargoLockPackageTransformer();
+    void transformNoVersion() {
+        CargoDependencyLineParser cargoDependencyLineParser = new CargoDependencyLineParser();
+        CargoLockPackageDataTransformer transformer = new CargoLockPackageDataTransformer(cargoDependencyLineParser);
 
-        assertThrows(DetectableException.class, () -> cargoLockPackageTransformer.transformToGraph(input));
-    }
+        CargoLockPackageData cargoLockPackageData = new CargoLockPackageData("some-name", null, null, null, Collections.emptyList());
+        CargoLockPackage cargoLockPackage = transformer.transform(cargoLockPackageData);
 
-    @Test
-    public void testCorrectNumberOfRootDependencies() throws DetectableException, MissingExternalIdException, CycleDetectedException {
-        List<CargoLockPackage> input = new ArrayList<>();
-        input.add(createPackage("test1", "1.0.0",
-            new NameOptionalVersion("dep1"),
-            new NameOptionalVersion("dep2")
-        ));
-        input.add(createPackage("dep1", "0.5.0"));
-        input.add(createPackage("dep2", "0.6.0"));
-        CargoLockPackageTransformer cargoLockPackageTransformer = new CargoLockPackageTransformer();
-        DependencyGraph graph = cargoLockPackageTransformer.transformToGraph(input);
-
-        NameVersionGraphAssert graphAssert = new NameVersionGraphAssert(Forge.CRATES, graph);
-        graphAssert.hasRootSize(1);
-    }
-
-    private CargoLockPackage createPackage(String name, String version, NameOptionalVersion... dependencies) {
-        return new CargoLockPackage(new NameVersion(name, version), Arrays.asList(dependencies));
+        assertEquals(new NameVersion("some-name", null), cargoLockPackage.getPackageNameVersion());
+        assertTrue(cargoLockPackage.getDependencies().isEmpty());
     }
 }
